Refactor: 상수 변경 및 Commands interface 상수파일로 분리
