+++ src/utils/constants.ts
@@ -1,5 +1,4 @@
 // export common interface and constants
-
 export interface GameState {
   WAITING: number;
   PROCEEDING: number;
@@ -16,8 +15,22 @@ export const GAME_STATE: GameState = {
   SETTING: 5,
 };
 
+export interface Commands {
+  command1: string;
+  command2: string;
+  command3: string;
+  command4: string;
+}
+
+export const INIT_COMMANDS: Commands = {
+  command1: "ArrowLeft",
+  command2: "ArrowUp",
+  command3: "ArrowRight",
+  command4: "Space" || "Spacebar",
+};
+
 export const MAX_TIMER = 90;
 export const MAX_BLOCK_LENGTH = 5;
 export const MAX_FEVER = 100;
 export const FEVER_UNIT = 10;
-export const FEVER_TIME_UNIT = 1000;
+export const FEVER_TIME_UNIT = 500;
